Submissions-2/E_Wallet/Group-6/testcodes/TransactionTest.java
import static org.junit.Assert.*;

import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;


public class TransactionTest {

	Transaction obj;
	@BeforeClass
	public static void setUpBeforeClass() throws Exception {
	}

	@AfterClass
	public static void tearDownAfterClass() throws Exception {
	}

	@Before
	public void setUp() throws Exception {
		obj = new Transaction(2500);
	}

	@After
	public void tearDown() throws Exception {
	}

	@Test
	public void test() {
		assertNotNull(obj);
		assertEquals( 2500,obj.printAmount());
		fail("Not yet implemented");
		
	}

}

